Add LocalConnectionProtocol object
Add a connection protocol that simply reflects its input back into its
output queue. Intended use is to derive a Connection object from it and
add it to the Client pool of the IListener object handling external
communications so that it is just like all of the other separate
connections.